jwtAuthCheck Middleware
